Add global version.  Update commentary and metadata.  Remove dash.
I updated the metadata for checkdoc and package-lint-current-buffer (required to submit to
MELPA).  This included adding a version which I've set to 2.0.0.  The previous keywords were
flagged as invalid, so I chose "convenience" from the list of approved (?) keywords.

I also merged in the extended commentary from my smart-tab-over package this is replacing to
explain *why* this is useful for those that are browsing packages and aren't necessarily
looking for this one.

I updated the list of characters to include > and :, but removed |.  The pipe doesn't seem like
an ending character.

Added tab-jump-out-global-mode, which is what I'll be using.

Removed the dependency on the dash library since it was only used for a single function which
matched the built-in `member` function.

I converted the readme from org to Markdown to simplify.  I'm personally a huge fan of org, but
this readme turned out much simpler without the #begin_src blocks.  (Plus I wasn't sure what
language to put for "M-x ..." ;)