Update recommendation of working with ya-snippet.
The suggestion to use yas-fallback-behavior no longer works so I've removed it.  It was marked
as obsolete quite a while back.  I've added an example of using the yas before and after hooks
which I use in my init.el.

I updated the version even though there was no code change so existing users would have the
docs available if they run into this problem.